+++ src/graphicals/sfml/Window/SFMLWindow.cpp
@@ -278,14 +278,14 @@ std::vector<shared::graphics::events::EventPtr> SFMLWindow::getEvents(void)
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonPressEvent>
                     (shared::graphics::events::IMouseButtonEvent::RIGHT,
-                    (Vector2i) {(event.mouseButton.x / static_cast<int>(sizeTiles.x)),
-                    (event.mouseButton.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.mouseButton.x / static_cast<float>(sizeTiles.x)),
+                    (event.mouseButton.y / static_cast<float>(sizeTiles.y))}));
                 } else if (event.mouseButton.button == sf::Mouse::Left) {
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonPressEvent>
                     (shared::graphics::events::IMouseButtonEvent::RIGHT,
-                    (Vector2i) {(event.mouseButton.x / static_cast<int>(sizeTiles.x)),
-                    (event.mouseButton.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.mouseButton.x / static_cast<float>(sizeTiles.x)),
+                    (event.mouseButton.y / static_cast<float>(sizeTiles.y))}));
                 }
                 break;
             case sf::Event::MouseButtonReleased:
@@ -293,20 +293,20 @@ std::vector<shared::graphics::events::EventPtr> SFMLWindow::getEvents(void)
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonReleaseEvent>
                     (shared::graphics::events::IMouseButtonEvent::LEFT,
-                    (Vector2i) {(event.mouseButton.x / static_cast<int>(sizeTiles.x)),
-                    (event.mouseButton.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.mouseButton.x / static_cast<float>(sizeTiles.x)),
+                    (event.mouseButton.y / static_cast<float>(sizeTiles.y))}));
                 } else if (event.mouseButton.button == sf::Mouse::Left) {
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonReleaseEvent>
                     (shared::graphics::events::IMouseButtonEvent::RIGHT,
-                    (Vector2i) {(event.mouseButton.x / static_cast<int>(sizeTiles.x)),
-                    (event.mouseButton.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.mouseButton.x / static_cast<float>(sizeTiles.x)),
+                    (event.mouseButton.y / static_cast<float>(sizeTiles.y))}));
                 }
                 break;
             case sf::Event::MouseMoved:
                 eventsList.push_back(std::make_shared<shared::graphics::events::MouseMoveEvent>
-                    ((Vector2i) {(event.mouseButton.x / static_cast<int>(sizeTiles.x)),
-                    (event.mouseButton.y / static_cast<int>(sizeTiles.y))}));
+                    ((Vector2f) {(event.mouseButton.x / static_cast<float>(sizeTiles.x)),
+                    (event.mouseButton.y / static_cast<float>(sizeTiles.y))}));
                 break;
             case sf::Event::KeyPressed:
                 KeyFindType = mapSFMLKeyToKeyType(event.key.code);
